moved get groups by profile functions to group functions file, added get games by group to game function file
