+++ src/Groups/groupsFunctions.js
@@ -1,4 +1,4 @@
-const {Group} = require("../database/schemas/groupsSchema")
+const { Group } = require("../database/schemas/groupsSchema")
 
 async function getAllGroups() {
     let allGroups = await Group.find()
@@ -26,7 +26,7 @@ async function updateSpecificGroup(groupId, requestBody) {
     console.log(requestBody)
     let updatedGroup = await Group.updateOne(
         // find the group with the Group ID in the query data
-        {_id: groupId},
+        { _id: groupId },
         // update/set the group object with all the query data
         { $set: requestBody }
     )
@@ -34,9 +34,25 @@ async function updateSpecificGroup(groupId, requestBody) {
 }
 
 async function deleteGroup(groupId) {
-    let deletedGroup = await Group.deleteOne({"_id": groupId})
+    let deletedGroup = await Group.deleteOne({ "_id": groupId })
     return deletedGroup
 }
 
+// returns array of groups where the input profile is a member.
+async function getJoinedGroupsByProfile(profileID) {
+    let profile = await getSpecificProfile(profileID)
+    let id = profile._id
+    let profileJoinedGroups = await Group.find({ members: id }); // Find Groups where "members" includes the specific profile
+    return profileJoinedGroups
+}
 
-module.exports = {getAllGroups, createNewGroup, getSpecificGroup, updateSpecificGroup, deleteGroup} 
+// returns array of groups where input profile is the group admin
+async function getAdminGroupsByProfile(profileID) {
+    let profile = await getSpecificProfile(profileID)
+    let id = profile._id
+    let profileAdministeredGroups = await Group.find({ adminID: id });
+    return profileAdministeredGroups
+}
+
+
+module.exports = { getAllGroups, createNewGroup, getSpecificGroup, updateSpecificGroup, deleteGroup }
