+++ src/Profiles/profilesFunctions.js
@@ -67,21 +67,6 @@ async function getGamesByProfile(profileID) {
   return profileGames;
 }
 
-// returns array of groups where the input profile is a member.
-async function getJoinedGroupsByProfile(profileID) {
-  let profile = await getSpecificProfile(profileID)
-  let id = profile._id
-  let profileJoinedGroups = await Group.find({ members: id }); // Find Groups where "members" includes the specific profile
-  return profileJoinedGroups
-}
-
-// returns array of groups where input profile is the group admin
-async function getAdminGroupsByProfile(profileID) {
-  let profile = await getSpecificProfile(profileID)
-  let id = profile._id
-  let profileAdministeredGroups = await Group.find({ adminID: id });
-  return profileAdministeredGroups
-}
 
 async function signInUser(queryData) {
   const firebaseClientAuth = getAuth();
